% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FluxtowerSpreadsheetToNc.R
\name{add_processing_info}
\alias{add_processing_info}
\title{Writes attributes common to met and flux NC files}
\usage{
add_processing_info(ncid, arg_info, datain, cat)
}
\description{
Writes attributes common to met and flux NC files
}

