% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Site_metadata.R
\name{add_processing_metadata}
\alias{add_processing_metadata}
\title{Adds processor metadata, including processor version}
\usage{
add_processing_metadata(metadata)
}
\value{
metadata list
}
\description{
Adds processor metadata, including processor version
}

