% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Aggregation.R
\name{aggregate_tsteps}
\alias{aggregate_tsteps}
\title{Aggregates original time resolution to a coarser time step}
\usage{
aggregate_tsteps(datain, new_tstep, qc_flags, qc_name)
}
\description{
Aggregates original time resolution to a coarser time step
}

