% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting.R
\name{AnnualCycle}
\alias{AnnualCycle}
\title{Plots an annual cycle}
\usage{
AnnualCycle(obslabel, acdata, varname, ytext, legendtext, timestepsize, whole,
  plotcolours, modlabel = "no", na.rm = FALSE)
}
\description{
Plots an annual cycle
}
