% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility_functions.R
\name{append_and_warn}
\alias{append_and_warn}
\title{Appends warning message and calls warning}
\usage{
append_and_warn(warn, warnings, call = TRUE)
}
\description{
Appends warning message and calls warning
}
