% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gapfilling.R
\name{calc_avPrecip}
\alias{calc_avPrecip}
\title{Calculates mean annual precipitation}
\usage{
calc_avPrecip(datain, gaps)
}
\arguments{
\item{datain}{input data}

\item{gaps}{gaps?}
}
\value{
mean annual preciptitation
}
\description{
Calculates mean annual precipitation
}
