% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Conversions.R
\name{calc_esat}
\alias{calc_esat}
\title{Calculates saturation vapour pressure}
\usage{
calc_esat(airtemp)
}
\arguments{
\item{airtemp}{air temperature in degrees C}
}
\value{
saturation vapour pressure
}
\description{
Calculates saturation vapour pressure
}
\references{
Jones (1992), Plants and microclimate: A quantitative approach to
environmental plant physiology, p110
}
