% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Conversions.R
\name{ChangeUnits}
\alias{ChangeUnits}
\title{Converts units from original Fluxnet to target ALMA units}
\usage{
ChangeUnits(datain)
}
\value{
datain
}
\description{
Converts units from original Fluxnet to target ALMA units
}

