% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility_functions.R
\name{check_flx2015_version}
\alias{check_flx2015_version}
\title{Checks  that FLUXNET2015 version defined correctly}
\usage{
check_flx2015_version(dataset, version)
}
\description{
Checks  that FLUXNET2015 version defined correctly
}
