% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Site_metadata.R
\name{check_missing}
\alias{check_missing}
\title{Checks which metadata are missing (correcting for OK NAs)}
\usage{
check_missing(metadata)
}
\value{
boolean metadata availability vector
}
\description{
Checks which metadata are missing (correcting for OK NAs)
}

