% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Check_and_Gapfill.R
\name{CheckDataGaps}
\alias{CheckDataGaps}
\title{Checks for data gaps and determines output years}
\usage{
CheckDataGaps(datain, missing_val = SprdMissingVal, threshold, min_yrs,
  essential_met, preferred_eval)
}
\value{
out
}
\description{
Checks for data gaps and determines output years
}

