% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Checks.R
\name{CheckDataRanges}
\alias{CheckDataRanges}
\title{Checks that data are within specified ranges
as set in the "variables" auxiliary file}
\usage{
CheckDataRanges(datain, site_log, action = "stop")
}
\arguments{
\item{datain}{input data}

\item{site_log}{log file}

\item{action}{action to execute upon failure?}
}
\value{
out
}
\description{
Checks that data are within specified ranges
as set in the "variables" auxiliary file
}
