% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Timing_general.R
\name{CheckTiming}
\alias{CheckTiming}
\title{Checks that have whole number of days in dataset}
\usage{
CheckTiming(datain, site_log)
}
\description{
Checks that have whole number of days in dataset
}

