% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FluxtowerSpreadsheetToNc.R
\name{convert_LaThuile}
\alias{convert_LaThuile}
\title{Converts La Thuile files to FLUXNET2015 format}
\usage{
convert_LaThuile(infiles, fair_use = NA, fair_use_vec = NA, min_yrs, tcol,
  site_log, ...)
}
\description{
Converts La Thuile files to FLUXNET2015 format
}

