% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Timing_general.R
\name{convert_LaThuile_time}
\alias{convert_LaThuile_time}
\title{Converts La Thuile time format to Fluxnet2015 time format}
\usage{
convert_LaThuile_time(timestep)
}
\description{
Converts La Thuile time format to Fluxnet2015 time format
}

