% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Handlers_csv.R
\name{convert_OzFlux}
\alias{convert_OzFlux}
\title{Converts OzFlux files to FLUXNET2015 format}
\usage{
convert_OzFlux(infile, tcol)
}
\arguments{
\item{infile}{input file}

\item{tcol}{tcol??}
}
\value{
input file
}
\description{
Converts OzFlux files to FLUXNET2015 format
}
