% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Timing_general.R
\name{convertToHoursMins}
\alias{convertToHoursMins}
\title{Utility function to convert minutes to HHMM format}
\usage{
convertToHoursMins(time, format = "\%02d\%02d")
}
\description{
Utility function to convert minutes to HHMM format
}

