% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Synthesize_and_gapfill_functions.R
\name{copyfill_data}
\alias{copyfill_data}
\title{Performs copyfill gapfilling}
\usage{
copyfill_data(data, tsteps, tstepsize, copyfill = 10, start, end, varname,
  site_log)
}
\description{
Performs copyfill gapfilling
}

