% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Timing_general.R
\name{create_dummy_year}
\alias{create_dummy_year}
\title{Function for filling missing years in La Thuile dataframe}
\usage{
create_dummy_year(year, tstep, time_vars)
}
\description{
Function for filling missing years in La Thuile dataframe
}

