% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Utility_functions.R
\name{create_outdir}
\alias{create_outdir}
\title{Creates output directories and returns log output path}
\usage{
create_outdir(outdir, site, plots)
}
\description{
Creates output directories and returns log output path
}

