% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gapfilling.R
\name{create_qc_var}
\alias{create_qc_var}
\title{Creates attributes for a new QC variable}
\usage{
create_qc_var(datain, qc_name, qc_flags, outname, cat)
}
\arguments{
\item{datain}{input data}

\item{qc_name}{qc name}

\item{qc_flags}{qc flag}

\item{outname}{outname}

\item{cat}{category?}
}
\value{
input data frame
}
\description{
Creates attributes for a new QC variable
}
