% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Handlers_NetCDF.R
\name{CreateMetNetcdfFile}
\alias{CreateMetNetcdfFile}
\title{Creates a netcdf file for met variables}
\usage{
CreateMetNetcdfFile(metfilename, datain, site_code, siteInfo, ind_start,
  ind_end, starttime, flux_varname, cf_name, av_precip = NA, total_missing,
  total_gapfilled, qcInfo, arg_info, var_ind, varnames, modelInfo)
}
\description{
Creates a netcdf file for met variables
}
