% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Timing_general.R
\name{DayNight}
\alias{DayNight}
\title{Finds day and night time steps}
\usage{
DayNight(SWdown, units)
}
\arguments{
\item{SWdown}{A numeric vector, surface incident shortwave radiation}

\item{units}{Character scalar, "umol/m2/s" or others}
}
\value{
\code{daynotnight}: A boolean vector. If \verb{Rs > 5 W/m2/s}, \code{daynotnight = TRUE},
else \code{daynotnight = FALSE}.
}
\description{
Finds day and night time steps
}
\references{
Adapted from PALS
}
