% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FluxtowerSpreadsheetToNc.R
\name{define_model_params}
\alias{define_model_params}
\title{Writes model parameters as global attribute to NetCDF file}
\usage{
define_model_params(modelInfo, dims)
}
\description{
Writes model parameters as global attribute to NetCDF file
}

