% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting.R
\name{DiurnalCycle}
\alias{DiurnalCycle}
\title{Plots a diurnal cycle}
\usage{
DiurnalCycle(
  obslabel,
  dcdata,
  varname,
  ytext,
  legendtext,
  timestepsize,
  whole,
  plotcolours,
  modlabel = "no",
  vqcdata = matrix(-1, nrow = 1, ncol = 1),
  plot.cex,
  na.rm = FALSE
)
}
\description{
Plots a diurnal cycle
}
