% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Check_and_Gapfill.R
\name{fill_qcvar_missing}
\alias{fill_qcvar_missing}
\title{Fills QC flags with 3 (poor gap-filling) when
QC flag missing but data variable available}
\usage{
fill_qcvar_missing(datain, missingVal, gapfillVal)
}
\value{
datain
}
\description{
Fills QC flags with 3 (poor gap-filling) when
QC flag missing but data variable available
}

