% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gapfilling.R
\name{FillQCvarMissing}
\alias{FillQCvarMissing}
\title{Fills QC flags}
\usage{
FillQCvarMissing(datain, gapfillVal, qc_name)
}
\arguments{
\item{datain}{input data}

\item{gapfillVal}{gapfilling value}

\item{qc_name}{qc name?}
}
\value{
gapfilled qc flags
}
\description{
with 3 (poor gap-filling) when
QC flag missing but data variable available
}
