% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Check_and_Gapfill.R
\name{find_exclude_eval}
\alias{find_exclude_eval}
\title{Finds evaluation variables to exlude}
\usage{
find_exclude_eval(datain, all_missing)
}
\description{
Finds evaluation variables to exlude
}

