% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Check_and_Gapfill.R
\name{find_flux_ind}
\alias{find_flux_ind}
\title{Finds indices for flux variables to be outputted}
\usage{
find_flux_ind(datain, exclude_eval, k, site_log)
}
\description{
Finds indices for flux variables to be outputted
}

