% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gapfilling.R
\name{find_ind_and_qc}
\alias{find_ind_and_qc}
\title{Find index for a variable and its QC flag}
\usage{
find_ind_and_qc(inds, var, qc_name = NA)
}
\arguments{
\item{inds}{indices}

\item{var}{variables}

\item{qc_name}{qc names}
}
\value{
indices for variables and QC flags
}
\description{
Find index for a variable and its QC flag
}
