% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Constants.R
\name{findColIndices}
\alias{findColIndices}
\title{Find list of column variables}
\usage{
findColIndices(
  fileinname,
  var_names,
  var_classes,
  essential_vars,
  preferred_vars,
  time_vars,
  dset_vars,
  site_log,
  datasetname,
  ...
)
}
\arguments{
\item{fileinname}{input file}

\item{var_names}{varnames}

\item{var_classes}{var classes}

\item{essential_vars}{essential variables}

\item{preferred_vars}{preferred variables}

\item{time_vars}{time step}

\item{dset_vars}{dataset variables?}

\item{site_log}{log file}

\item{datasetname}{dataset name}

\item{...}{additional parameters}
}
\value{
list of variables and their attributes
}
\description{
Process files for their column variables
}
