% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Gapfilling.R
\name{FindExcludeEval}
\alias{FindExcludeEval}
\title{Finds evaluation variables to exlude}
\usage{
FindExcludeEval(datain, all_missing, gaps, include_all, qc_name)
}
\description{
Finds evaluation variables to exlude
}

