% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Timing_general.R
\name{findStartTime}
\alias{findStartTime}
\title{Timing_general.R}
\usage{
findStartTime(start)
}
\description{
Functions that assess or convert timing variables.
}
\details{
TODO: Check and merge back in to palsR
}
