% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Constants.R
\name{findTimeInfo}
\alias{findTimeInfo}
\title{Extract time stamp information}
\usage{
findTimeInfo(time_vars, headers, site_log, datasetname)
}
\arguments{
\item{time_vars}{time variables}

\item{headers}{headers}

\item{site_log}{log files}

\item{datasetname}{filename}
}
\value{
time stamp variables
}
\description{
Extract time stamp information
}
