% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fluxnetLSM-package.R
\docType{package}
\name{fluxnetLSM-package}
\alias{fluxnetLSM}
\alias{fluxnetLSM-package}
\title{fluxnetLSM: Quality control and data conversion scripts for processing FLUXNET data 
    for land surface modelling}
\description{
Scripts for quality control of FLUXNET data, and converting from 
    CSV format to NetCDF, and managing metadata.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/aukkola/FluxnetLSM}
}

}
\keyword{internal}
