\name{FluxnetProcessing-package}
\alias{FluxnetProcessing-package}
\alias{FluxnetProcessing}
\docType{package}
\title{
\packageTitle{FluxnetProcessing}
}
\description{
\packageDescription{FluxnetProcessing}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{FluxnetProcessing}
\packageIndices{FluxnetProcessing}
% TODO: An overview of how to use the package, including the most important functions
}
\author{
\packageAuthor{FluxnetProcessing}

Maintainer: \packageMaintainer{FluxnetProcessing}
}
\references{
% TODO: Literature or other references for background information 
}
% TODO: Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory
\keyword{ package }
\seealso{
% TODO: Optional links to other man pages, e.g.
% TODO: \code{\link[<pkg>:<pkg>-package]{<pkg>}}
}
\examples{
% TODO: simple examples of the most important functions
}
