% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Synthesize_and_gapfill_functions.R
\name{gapfill_LWdown_Pair}
\alias{gapfill_LWdown_Pair}
\title{Synthesizes incoming longwave radiation and air pressure}
\usage{
gapfill_LWdown_Pair(
  data,
  var,
  var_ind,
  TairK = NA,
  RH = NA,
  technique = NA,
  elev = NA,
  varnames,
  site_log
)
}
\description{
Synthesizes incoming longwave radiation and air pressure
}
