% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Synthesize_and_gapfill_functions.R
\name{gapfill_with_ERA}
\alias{gapfill_with_ERA}
\title{Gapfills met data}
\usage{
gapfill_with_ERA(
  datain,
  era_data,
  era_vars,
  tair_units,
  vpd_units,
  out_vars,
  qc_name,
  qc_flags,
  varnames,
  site_log
)
}
\value{
out
}
\description{
Gapfills met data
}
