% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Check_and_Gapfill.R
\name{GapfillFlux}
\alias{GapfillFlux}
\title{Gapfill flux variables using statistical methods}
\usage{
GapfillFlux(datain, qc_name, qc_flags, regfill, linfill, copyfill, gaps,
  site_log)
}
\description{
Gapfill flux variables using statistical methods
}

