% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Check_and_Gapfill.R
\name{GapfillMet}
\alias{GapfillMet}
\title{Gapfills met data}
\usage{
GapfillMet(datain, era_data, era_vars, tair_units, vpd_units, missing_val,
  out_vars)
}
\value{
out
}
\description{
Gapfills met data
}

