% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Check_and_Gapfill.R
\name{GapfillMet_statistical}
\alias{GapfillMet_statistical}
\title{Gapfill meteorological variables using statistical methods}
\usage{
GapfillMet_statistical(datain, qc_name, qc_flags, copyfill, linfill,
  lwdown_method, elevation, gaps, site_log)
}
\description{
Gapfill meteorological variables using statistical methods
}

