% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Site_metadata.R
\name{get_fluxdata_org_site_metadata}
\alias{get_fluxdata_org_site_metadata}
\title{Get metadata from Fluxdata.org}
\usage{
get_fluxdata_org_site_metadata(metadata, site_url = NULL)
}
\value{
metadata list
}
\description{
Get metadata from Fluxdata.org
}
