% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helpers.R
\name{get_fluxnet_erai_template}
\alias{get_fluxnet_erai_template}
\title{Gets a Fluxnet ERA interim file template (or full file name if site codes and years are specified)}
\usage{
get_fluxnet_erai_template(site_code, ...)
}
\description{
Gets a Fluxnet ERA interim file template (or full file name if site codes and years are specified)
}
