% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Helpers.R
\name{get_fluxnet_file_template}
\alias{get_fluxnet_file_template}
\title{Gets a Fluxnet file template (or full file name if site codes and years are specified)}
\usage{
get_fluxnet_file_template(site_code = "??-???", datasetname = "FLUXNET2015",
  subset = "FULLSET", resolution = "HH", years = "[0-9]{4}-[0-9]{4}",
  datasetversion = "1.3", extension = ".csv")
}
\description{
Gets a Fluxnet file template (or full file name if site codes and years are specified)
}

