% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Helpers.R
\name{get_fluxnet_files}
\alias{get_fluxnet_files}
\title{Gets Fluxnet file(s) available at a given path}
\usage{
get_fluxnet_files(path, site_code = "[A-Z]{2}-[A-Za-z0-9]{3}", ...)
}
\description{
Gets Fluxnet file(s) available at a given path
}

