% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Helpers.R
\name{get_fluxnet_site_code}
\alias{get_fluxnet_site_code}
\title{Gets Fluxnet site_code from Fluxnet file.
Useful when processing multiple files}
\usage{
get_fluxnet_site_code(path, file, ...)
}
\description{
Gets Fluxnet site_code from Fluxnet file.
Useful when processing multiple files
}

