% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Site_metadata.R
\name{get_git_version}
\alias{get_git_version}
\title{Gets the git version from the installed package
See src/zzz.R for how git revision is discovered}
\usage{
get_git_version()
}
\description{
Gets the git version from the installed package
See src/zzz.R for how git revision is discovered
}

