% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helpers.R
\name{get_lathuile_datapolicy}
\alias{get_lathuile_datapolicy}
\title{Gets La Thuile data policy for each site year}
\usage{
get_lathuile_datapolicy(site_code, site_use)
}
\description{
Gets La Thuile data policy for each site year
}
