% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Site_metadata.R
\name{get_ornl_site_metadata}
\alias{get_ornl_site_metadata}
\title{Get metadata from ORNL}
\usage{
get_ornl_site_metadata(metadata, site_url = NULL, overwrite = TRUE)
}
\value{
metadata list
}
\description{
Get metadata from ORNL
}

