% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Site_metadata.R
\name{get_ornl_site_url_list}
\alias{get_ornl_site_url_list}
\title{Get a list of ORNL site URLs from site_status table}
\usage{
get_ornl_site_url_list(site_code_list)
}
\description{
Get a list of ORNL site URLs from site_status table
}

