% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helpers.R, R/Site_metadata.R
\name{get_site_code}
\alias{get_site_code}
\alias{get_site_code}
\title{Gets Fluxnet site_code from Fluxnet file.
Useful when processing multiple files}
\usage{
get_site_code(metadata)

get_site_code(metadata)
}
\description{
Gets Fluxnet site_code from Fluxnet file.
Useful when processing multiple files

Get site name from metadata
}
