% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Site_metadata.R
\name{get_site_metadata}
\alias{get_site_metadata}
\title{Get site metadata. Tries multiple methods to retrieve full metadata}
\usage{
get_site_metadata(site_code, incl_processing = TRUE, use_csv = TRUE,
  update_csv = FALSE)
}
\value{
metadata list
}
\description{
Get site metadata. Tries multiple methods to retrieve full metadata
}
