% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Site_metadata.R
\name{get_site_metadata_CSV}
\alias{get_site_metadata_CSV}
\title{Tries to gather metadata from the included site CSV}
\usage{
get_site_metadata_CSV(metadata)
}
\value{
metadata list
}
\description{
Tries to gather metadata from the included site CSV
}

