% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Site_metadata.R
\name{get_site_metadata_web}
\alias{get_site_metadata_web}
\title{Tries to load metadata from known Fluxnet info sources on the 'web}
\usage{
get_site_metadata_web(metadata)
}
\value{
metadata list
}
\description{
Tries to load metadata from known Fluxnet info sources on the 'web
}

