% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility_functions.R
\name{get_varnames}
\alias{get_varnames}
\title{Gets possible varnames for FLUXNET FULLSET/SUBSET and La Thuile}
\usage{
get_varnames(datasetname, flx2015_version)
}
\description{
Gets possible varnames for FLUXNET FULLSET/SUBSET and La Thuile
}
