% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Checks.R
\name{InitialChecks}
\alias{InitialChecks}
\title{Performs initial checks on function arguments}
\usage{
InitialChecks(opts, era_file)
}
\arguments{
\item{opts}{options}

\item{era_file}{ERA file to process}
}
\description{
Performs initial checks on function arguments
}
