% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Checks.R
\name{IsWholeYrs}
\alias{IsWholeYrs}
\title{Checks that whole years were extracted}
\usage{
IsWholeYrs(datain, gaps, site_log)
}
\arguments{
\item{datain}{input data}

\item{gaps}{gaps}

\item{site_log}{log file}
}
\description{
Checks that whole years were extracted
}
