% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{LaThuile_site_policy}
\alias{LaThuile_site_policy}
\title{DATASET_TITLE}
\format{A data frame with 253 rows and 18 variables:
\describe{
\item{\code{V1}}{character COLUMN_DESCRIPTION}
\item{\code{V2}}{character COLUMN_DESCRIPTION}
\item{\code{V3}}{character COLUMN_DESCRIPTION}
\item{\code{V4}}{character COLUMN_DESCRIPTION}
\item{\code{V5}}{character COLUMN_DESCRIPTION}
\item{\code{V6}}{character COLUMN_DESCRIPTION}
\item{\code{V7}}{character COLUMN_DESCRIPTION}
\item{\code{V8}}{character COLUMN_DESCRIPTION}
\item{\code{V9}}{character COLUMN_DESCRIPTION}
\item{\code{V10}}{character COLUMN_DESCRIPTION}
\item{\code{V11}}{character COLUMN_DESCRIPTION}
\item{\code{V12}}{character COLUMN_DESCRIPTION}
\item{\code{V13}}{character COLUMN_DESCRIPTION}
\item{\code{V14}}{character COLUMN_DESCRIPTION}
\item{\code{V15}}{character COLUMN_DESCRIPTION}
\item{\code{V16}}{character COLUMN_DESCRIPTION}
\item{\code{V17}}{character COLUMN_DESCRIPTION}
\item{\code{V18}}{character COLUMN_DESCRIPTION}
}}
\source{
\url{http://somewhere.important.com/}
}
\usage{
LaThuile_site_policy
}
\description{
DATASET_DESCRIPTION
}
\keyword{datasets}
