% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Synthesize_and_gapfill_functions.R
\name{linfill_data}
\alias{linfill_data}
\title{Performs linear interpolation gapfilling}
\usage{
linfill_data(data, tstepsize, linfill = 10)
}
\description{
Performs linear interpolation gapfilling
}

