% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Site_metadata.R
\name{metadata_list_to_dataframe}
\alias{metadata_list_to_dataframe}
\title{Convert a list of metadata lists to a dataframe}
\usage{
metadata_list_to_dataframe(metadata_lists)
}
\description{
Convert a list of metadata lists to a dataframe
}

