% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Output_variables_FLUXNET2015_FULLSET}
\alias{Output_variables_FLUXNET2015_FULLSET}
\title{Output_variables_FLUXNET2015_FULLSET}
\format{A data frame with 40 rows and 15 variables:
\describe{
\item{\code{Fluxnet_variable}}{character COLUMN_DESCRIPTION}
\item{\code{Fluxnet_unit}}{character COLUMN_DESCRIPTION}
\item{\code{Fluxnet_class}}{character COLUMN_DESCRIPTION}
\item{\code{Output_variable}}{character COLUMN_DESCRIPTION}
\item{\code{Output_unit}}{character COLUMN_DESCRIPTION}
\item{\code{Longname}}{character COLUMN_DESCRIPTION}
\item{\code{Standard_name}}{character COLUMN_DESCRIPTION}
\item{\code{CMIP_short_name}}{character COLUMN_DESCRIPTION}
\item{\code{Data_min}}{integer COLUMN_DESCRIPTION}
\item{\code{Data_max}}{double COLUMN_DESCRIPTION}
\item{\code{Essential_met}}{integer COLUMN_DESCRIPTION}
\item{\code{Preferred_eval}}{logical COLUMN_DESCRIPTION}
\item{\code{Category}}{character COLUMN_DESCRIPTION}
\item{\code{ERAinterim_variable}}{character COLUMN_DESCRIPTION}
\item{\code{Aggregate_method}}{character COLUMN_DESCRIPTION}
}}
\source{
\url{http://somewhere.important.com/}
}
\usage{
Output_variables_FLUXNET2015_FULLSET
}
\description{
DATASET_DESCRIPTION
}
\keyword{datasets}
