% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting.R
\name{plot_nc}
\alias{plot_nc}
\title{Plots standard analysis plots from netcdf data}
\usage{
plot_nc(
  ncfile,
  analysis_type,
  vars,
  varnames = NULL,
  outfile = "fluxnetLSM_",
  qc_flags
)
}
\arguments{
\item{ncfile}{an open netcdf file}

\item{analysis_type}{vector of plot names: c("annual", "diurnal", "timeseries")}

\item{vars}{vector of variable names to plot}

\item{varnames}{(optional) invisible for users}

\item{outfile}{(optional) output path prefix, including directory}

\item{qc_flags}{(optional) list of \code{QC_measured}, \code{QC_gapfilled} and \code{qc_info}}
}
\description{
Plots standard analysis plots from netcdf data
}
\examples{
\dontrun{
library(ncdf4)
library(magrittr)

file   = "OUTPUTS/Nc_files/Flux/AU-Whr_2012-2014_FLUXNET2015_Flux.nc"
ncfile = nc_open(file)

vars_ndim <- map_int(ncfile$var, ~length(.x$varsize))
vars <- vars_ndim \%>\% {names(.)[. == 3]}

plot_nc(ncfile, analysis_type = c("diurnal"), vars, outfile = "diurnal_")
# SumatraPDF("diurnal_DiurnalCycle.pdf")
}
}
