% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Aggregation.R
\name{qc_frac}
\alias{qc_frac}
\title{Calculates fraction of good quality data}
\usage{
qc_frac(data, good_data)
}
\arguments{
\item{data}{original input data}

\item{good_data}{gapfilled data}
}
\description{
Returns gapfilled fraction given two data frames
of original and gappfilled data
}
