% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Handlers_csv.R
\name{read_era}
\alias{read_era}
\title{Reads ERA data and extracts time steps corresponding to obs}
\usage{
read_era(ERA_file, datain)
}
\arguments{
\item{ERA_file}{ERA5 file}

\item{datain}{input data file}
}
\value{
extracted values
}
\description{
Reads ERA data and extracts time steps corresponding to obs
}
