% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Handlers_csv.R
\name{ReadCSVFluxData}
\alias{ReadCSVFluxData}
\title{Reads comma-delimited text files containing Fluxnet2015 data}
\usage{
ReadCSVFluxData(
  fileinname,
  vars,
  datasetname,
  time_vars,
  add_psurf,
  site_log,
  ...
)
}
\arguments{
\item{fileinname}{fluxnet data file, including directory}

\item{vars}{data.table of variables and their attributes}

\item{time_vars}{vector of time variables}
}
\value{
list of flux data, variables and timing information
}
\description{
Reads comma-delimited text files containing Fluxnet2015 data
}
