% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FluxtowerSpreadsheetToNc.R
\name{ReadTextFluxData}
\alias{ReadTextFluxData}
\title{Reads comma-delimited text files containing Fluxnet2015 data}
\usage{
ReadTextFluxData(fileinname, vars, time_vars)
}
\arguments{
\item{fileinname}{fluxnet data file, including directory}

\item{vars}{data.table of variables and their attributes}

\item{time_vars}{vector of time variables}
}
\value{
list of flux data, variables and timing information
}
\description{
Reads comma-delimited text files containing Fluxnet2015 data
}

