% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Synthesize_and_gapfill_functions.R
\name{regpredict}
\alias{regpredict}
\title{Predict flux values using linear regression parameters}
\usage{
regpredict(rgrp, traindata, dayn)
}
\description{
Predict flux values using linear regression parameters
}

