% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Conversions.R
\name{Rel2SpecHum}
\alias{Rel2SpecHum}
\title{Converts relative humidity to specific humidity.}
\usage{
Rel2SpecHum(relHum, airtemp, tair_units, pressure, psurf_units)
}
\value{
specific humidity in kg/kg
}
\description{
Converts relative humidity to specific humidity.
}

