% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Conversions.R
\name{Rel2SpecHumidity}
\alias{Rel2SpecHumidity}
\title{Converts relative humidity to specific humidity.}
\usage{
Rel2SpecHumidity(relHum, airtemp, tair_units, pressure, psurf_units, site_log)
}
\value{
specific humidity in kg/kg
}
\description{
Converts relative humidity to specific humidity.
}

