% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Constants.R
\name{remove_duplicates}
\alias{remove_duplicates}
\title{Removes duplicate indices}
\usage{
remove_duplicates(indices)
}
\arguments{
\item{indices}{indices to remove}
}
\value{

}
\description{
If a Fluxnet variable is processed more than once remove doubless
}
