% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Constants.R
\name{rename_vars}
\alias{rename_vars}
\title{Renames Fluxnet variables to desired output variable names}
\usage{
rename_vars(vars_present, all_vars)
}
\value{
renamed variables
}
\description{
Renames Fluxnet variables to desired output variable names
}
