% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Constants.R
\name{retrieve_categories}
\alias{retrieve_categories}
\title{Retrieves variable categories to divide into met and flux data}
\usage{
retrieve_categories(vars_present, all_vars)
}
\value{
vat_vars
}
\description{
Retrieves variable categories to divide into met and flux data
}

