% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Constants.R
\name{retrieve_outnames}
\alias{retrieve_outnames}
\title{Retrieves original and target variable units}
\usage{
retrieve_outnames(vars_present, all_vars)
}
\value{
original and target units
}
\description{
Retrieves original and target variable units
}

