% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Constants.R
\name{retrieve_ranges}
\alias{retrieve_ranges}
\title{Retrieves acceptable variable ranges}
\usage{
retrieve_ranges(vars_present, all_vars)
}
\arguments{
\item{vars_present}{current variable ranges}

\item{all_vars}{all variable ranges}
}
\value{

}
\description{
Retrieves acceptable variable ranges
}
