% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Constants.R
\name{retrieve_units}
\alias{retrieve_units}
\title{Retrieves original and target variable units}
\usage{
retrieve_units(vars_present, all_vars)
}
\value{
original and target units
}
\description{
Retrieves original and target variable units
}

