% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Site_metadata.R
\name{save_metadata_list_to_csv}
\alias{save_metadata_list_to_csv}
\title{Write multiple site metadata to list at once}
\usage{
save_metadata_list_to_csv(metadata_lists)
}
\description{
Write multiple site metadata to list at once
}

