% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Site_metadata.R
\name{save_metadata_to_csv}
\alias{save_metadata_to_csv}
\title{Writes metadata to CSV, only updating non-NA data}
\usage{
save_metadata_to_csv(metadata)
}
\description{
Writes metadata to CSV, only updating non-NA data
}
