% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{site_metadata}
\alias{site_metadata}
\title{DATASET_TITLE}
\format{A data frame with 856 rows and 22 variables:
\describe{
\item{\code{SiteCode}}{character COLUMN_DESCRIPTION}
\item{\code{Exclude}}{character COLUMN_DESCRIPTION}
\item{\code{Exclude_reason}}{logical COLUMN_DESCRIPTION}
\item{\code{CABLE_PFT}}{integer COLUMN_DESCRIPTION}
\item{\code{Description}}{character COLUMN_DESCRIPTION}
\item{\code{TowerStatus}}{character COLUMN_DESCRIPTION}
\item{\code{Country}}{character COLUMN_DESCRIPTION}
\item{\code{Fullname}}{character COLUMN_DESCRIPTION}
\item{\code{SiteLatitude}}{double COLUMN_DESCRIPTION}
\item{\code{SiteLongitude}}{double COLUMN_DESCRIPTION}
\item{\code{SiteElevation}}{double COLUMN_DESCRIPTION}
\item{\code{IGBP_vegetation_short}}{character COLUMN_DESCRIPTION}
\item{\code{IGBP_vegetation_long}}{character COLUMN_DESCRIPTION}
\item{\code{Tier}}{integer COLUMN_DESCRIPTION}
\item{\code{MeasurementHeight}}{double COLUMN_DESCRIPTION}
\item{\code{TowerHeight}}{character COLUMN_DESCRIPTION}
\item{\code{CanopyHeight}}{double COLUMN_DESCRIPTION}
\item{\code{VegetationDescription}}{character COLUMN_DESCRIPTION}
\item{\code{SoilType}}{character COLUMN_DESCRIPTION}
\item{\code{Disturbance}}{character COLUMN_DESCRIPTION}
\item{\code{CropDescription}}{character COLUMN_DESCRIPTION}
\item{\code{Irrigation}}{character COLUMN_DESCRIPTION}
}}
\source{
\url{http://somewhere.important.com/}
}
\usage{
site_metadata
}
\description{
DATASET_DESCRIPTION
}
\keyword{datasets}
