% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Site_metadata.R
\name{site_metadata_template}
\alias{site_metadata_template}
\title{Empty site metadata template, including only the site code}
\usage{
site_metadata_template(site_code)
}
\value{
metadata list
}
\description{
Empty site metadata template, including only the site code
}
