% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Constants.R
\docType{data}
\name{SprdMissingVal}
\alias{SprdMissingVal}
\title{Constants.R}
\format{\preformatted{ num -9999
}}
\usage{
SprdMissingVal
}
\description{
Functions to find variables and their attributes
from Fluxnet csv-file
}
\details{
TODO: Check and merge back in to palsR
}
\keyword{datasets}

