% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Utility_functions.R
\name{stop_and_log}
\alias{stop_and_log}
\title{Writes site log and then aborts, reporting error}
\usage{
stop_and_log(error, site_log)
}
\description{
Writes site log and then aborts, reporting error
}

