% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Synthesize_and_gapfill_functions.R
\name{SynthesizeLWdown}
\alias{SynthesizeLWdown}
\title{Synthesises downward longwave radiation based on Tair and rel humidity}
\usage{
SynthesizeLWdown(TairK, RH, technique)
}
\description{
Synthesises downward longwave radiation based on Tair and rel humidity
}

