% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Synthesize_and_gapfill_functions.R
\name{SynthesizePSurf}
\alias{SynthesizePSurf}
\title{Synthesises air pressure based on Tair and elevation}
\usage{
SynthesizePSurf(TairK, elevation, pair_units)
}
\description{
Synthesises air pressure based on Tair and elevation
}
