% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Plotting.R
\name{Timeseries}
\alias{Timeseries}
\title{Plots a smoothed and non-smoothed time seris}
\usage{
Timeseries(obslabel, tsdata, varname, ytext, legendtext, plotcex, timing,
  smoothed = FALSE, winsize = 1, plotcolours, modlabel = "no",
  vqcdata = matrix(-1, nrow = 1, ncol = 1), na.rm = FALSE)
}
\description{
Plots a smoothed and non-smoothed time seris
}

