% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Site_metadata.R
\name{update_metadata}
\alias{update_metadata}
\title{Updates old metadata with new metadata, ignoring NAs, and warning on differences}
\usage{
update_metadata(metadata, new_metadata, overwrite = TRUE)
}
\value{
metadata list
}
\description{
Updates old metadata with new metadata, ignoring NAs, and warning on differences
}

