% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility_functions.R
\name{Utility}
\alias{Utility}
\alias{log_error}
\alias{log_warning}
\alias{write_log}
\title{Functions for error and warning handling}
\usage{
log_error(errtext, site_log)

log_warning(warn, site_log)

write_log(site_log)
}
\description{
\itemize{
\item \code{log_error}  : Append error message to log
\item \code{log_warning}: Appends warning to site log
\item \code{write_log}  : Writes site log
}
}
\keyword{internal}
