% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Utility_functions.R
\name{warn_and_log}
\alias{warn_and_log}
\title{Writes site log and then aborts, reporting error}
\usage{
warn_and_log(warn, site_log)
}
\description{
Writes site log and then aborts, reporting error
}

