% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Site_metadata.R
\name{warn_missing_metadata}
\alias{warn_missing_metadata}
\title{Warns about missing metadata for the site}
\usage{
warn_missing_metadata(metadata)
}
\description{
Warns about missing metadata for the site
}

