% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FluxtowerSpreadsheetToNc.R
\name{write_model_params}
\alias{write_model_params}
\title{Writes model parameters as global attribute to NetCDF file}
\usage{
write_model_params(ncid, modelInfo, missing_val)
}
\description{
Writes model parameters as global attribute to NetCDF file
}

