% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Handlers_NetCDF.R
\name{add_processing_info}
\alias{add_processing_info}
\title{Writes attributes common to met and flux NC files}
\usage{
add_processing_info(ncid, arg_info, datain, cat)
}
\arguments{
\item{ncid}{netcdf id}

\item{arg_info}{arguments}

\item{datain}{input data}

\item{cat}{category??}
}
\description{
Writes attributes common to met and flux NC files
}
