% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Conversions.R
\name{calc_esat}
\alias{calc_esat}
\title{Calculates saturation vapour pressure}
\usage{
calc_esat(airtemp)
}
\value{
saturation vapour pressure
}
\description{
Calculates saturation vapour pressure
}

