% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Check_and_Gapfill.R
\name{CheckDataGaps}
\alias{CheckDataGaps}
\title{Check_and_Gapfill.R}
\usage{
CheckDataGaps(datain, missing_val = SprdMissingVal, threshold, min_yrs,
  essential_met, preferred_eval)
}
\description{
A collections of functions for checking for
data gaps and ranges, and gap-filling
}
\details{
TODO: Check and merge back in to palsR
}

