% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Checks.R
\name{CheckDataGaps}
\alias{CheckDataGaps}
\title{Checks for missing and gap-filled data and determines output years}
\usage{
CheckDataGaps(
  datain,
  qc_flags,
  missing_met,
  missing_flux,
  gapfill_met_tier1,
  gapfill_met_tier2,
  gapfill_flux,
  gapfill_good,
  gapfill_med,
  gapfill_poor,
  gapfill_era,
  gapfill_stat,
  min_yrs,
  qc_name,
  aggregate = NA,
  site_log
)
}
\arguments{
\item{datain}{Input data list}

\item{qc_flags}{Values of qc flags}

\item{missing_met}{Threshold for missing values per year for met variables (as percentage)}

\item{missing_flux}{Threshold for missing values per year for flux variables (as percentage)}

\item{gapfill_met_tier1}{Threshold for all gap_filling per year for Tier 1 met variables (as percentage)}

\item{gapfill_met_tier2}{Threshold for all gap_filling per year for Tier 2 met variables (as percentage)}

\item{gapfill_flux}{Threshold for all gap_filling per year for flux variables (as percentage)}

\item{gapfill_good}{Threshold for good quality gap_filling
per year (as percentage, ignored if gapfill_all set)}

\item{gapfill_med}{Threshold for medium quality
gap_filling per year (as percentage, ignored if gapfill_all set)}

\item{gapfill_poor}{Threshold for poor quality gap_filling
per year (as percentage, ignored if gapfill_all set)}

\item{min_yrs}{Minimum number of consecutive years}

\item{qc_name}{Name of QC variable of dataset}

\item{site_log}{Site log}

\item{showWarn}{Print warning?}
}
\value{
out
}
\description{
Checks for missing and gap-filled data and determines output years
}
