% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Check_and_Gapfill.R
\name{CheckDataGaps}
\alias{CheckDataGaps}
\title{Checks for missing and gap-filled data and determines output years}
\usage{
CheckDataGaps(datain, missing_val, QCmeasured, QCgapfilled, missing,
  gapfill_all, gapfill_good, gapfill_med, gapfill_poor, min_yrs, essential_met,
  preferred_eval, all_eval)
}
\arguments{
\item{datain}{Input data list}

\item{missing_val}{Missing value}

\item{QCmeasured}{Value of measured qc flag}

\item{QCgapfilled}{Values of gap-filled qc flags}

\item{missing}{Threshold for missing values per year (as percentage)}

\item{gapfill_all}{Threshold for all gap_filling per year (as percentage)}

\item{gapfill_good}{Threshold for good quality gap_filling
per year (as percentage, ignored if gapfill_all set)}

\item{gapfill_med}{Threshold for medium quality
gap_filling per year (as percentage, ignored if gapfill_all set)}

\item{gapfill_poor}{Threshold for poor quality gap_filling
per year (as percentage, ignored if gapfill_all set)}

\item{min_yrs}{Minimum number of consecutive years}

\item{essential_met}{Names of essential met variables}

\item{preferred_eval}{Names of preferred evaluation variables}
}
\value{
out
}
\description{
Checks for missing and gap-filled data and determines output years
}

