% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Checks.R
\name{CheckDataGaps}
\alias{CheckDataGaps}
\title{Checks for missing and gap-filled data and determines output years}
\usage{
CheckDataGaps(datain, qc_flags, missing, gapfill_all, gapfill_good, gapfill_med,
  gapfill_poor, min_yrs, qc_name, showWarn = TRUE, aggregate = NA, site_log)
}
\arguments{
\item{datain}{Input data list}

\item{qc_flags}{Values of qc flags}

\item{missing}{Threshold for missing values per year (as percentage)}

\item{gapfill_all}{Threshold for all gap_filling per year (as percentage)}

\item{gapfill_good}{Threshold for good quality gap_filling 
per year (as percentage, ignored if gapfill_all set)}

\item{gapfill_med}{Threshold for medium quality 
gap_filling per year (as percentage, ignored if gapfill_all set)}

\item{gapfill_poor}{Threshold for poor quality gap_filling 
per year (as percentage, ignored if gapfill_all set)}

\item{min_yrs}{Minimum number of consecutive years}

\item{qc_name}{Name of QC variable of dataset}

\item{showWarn}{Print warning?}

\item{site_log}{Site log}
}
\value{
out
}
\description{
Checks for missing and gap-filled data and determines output years
}
