% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Checks.R
\name{CheckDataRanges}
\alias{CheckDataRanges}
\title{Checks that data are within specified ranges}
\usage{
CheckDataRanges(datain, site_log, action = "stop")
}
\description{
Checks that data are within specified ranges
}
