% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Check_and_Gapfill.R
\name{CheckDataRanges}
\alias{CheckDataRanges}
\title{Checks that data are within specified ranges}
\usage{
CheckDataRanges(datain, missingval, site_log)
}
\description{
Checks that data are within specified ranges
}

