% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConvertSpreadsheetToNcdf.R
\name{convert_fluxnet_to_netcdf}
\alias{convert_fluxnet_to_netcdf}
\title{ConvertSpreadsheetToNcdf.R}
\usage{
convert_fluxnet_to_netcdf(infile, site_code, out_path, lib_path,
  era_file = NA, era_gapfill = FALSE, datasetname = "Fluxnet2015",
  datasetversion = "Nov16", gap_threshold = 20, min_yrs = 2)
}
\arguments{
\item{infile}{input filename,
e.g. "FULLSET/FLX_AU-How_FLUXNET2015_FULLSET_HH_2001-2014_1-3.csv"}

\item{site_code}{Fluxnet site code e.g. "AU-How"}

\item{out_path}{output path e.g. "~/Documents/FLUXNET2016_processing/"}

\item{era_file}{ERA input file (needed if using ERAinterim to gapfill met variables)
e.g. "FULLSET/FLX_AU-How_FLUXNET2015_ERAI_HH_1989-2014_1-3.csv"}

\item{min_yrs}{Minimum numbers of years to process}

\item{threshold}{How many percent of time steps allowed to be missing in any given year?}

\item{ERA_gapfill}{Gapfill met variables using ERAinterim?}
}
\description{
Converts data from a PALS formatted spreadhseet to
netcdf.
}
\details{
Gab Abramowitz UNSW 2012 (palshelp at gmail dot com)
Main function to convert Fluxnet2015 CSV-files to NetCDF
}
