% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ConvertSpreadsheetToNcdf.R
\name{convert_fluxnet_to_netcdf}
\alias{convert_fluxnet_to_netcdf}
\title{ConvertSpreadsheetToNcdf.R}
\usage{
convert_fluxnet_to_netcdf(infile, site_code, out_path, ERA_file = NA,
  ERA_gapfill = FALSE, datasetname = "FLUXNET2015",
  datasetversion = "1-3", fair_use = NA, fair_use_vec = NA,
  missing = 15, gapfill_all = 20, gapfill_good = NA, gapfill_med = NA,
  gapfill_poor = NA, min_yrs = 2, include_all_eval = TRUE,
  plot = c("annual", "diurnal", "timeseries"))
}
\arguments{
\item{infile}{input filename,
e.g. "FULLSET/FLX_AU-How_FLUXNET2015_FULLSET_HH_2001-2014_1-3.csv"}

\item{site_code}{Fluxnet site code e.g. "AU-How"}

\item{out_path}{output path e.g. "./FLUXNET2016_processing/"}

\item{ERA_gapfill}{Gapfill met variables using ERAinterim?}

\item{datasetname}{Name of the dataset, e.g. FLUXNET2015}

\item{datasetversion}{Version of the dataset, e.g. "1-3"}

\item{fair_use}{Fair Use policy that data should comply with, e.g. "Tier1" for FLUXNET2015 or
"LaThuile" or "Fair_Use" for LaThuile Synthesis. Can be a single entry or vector of several policies.
If this is set, code will only extract years that comply with the required policy/policies. Must provide
fair_use_vec to use this functionality.}

\item{fair_use_vec}{A vector of Data Use policy for each year in the data file, e.g. "LaThuile" or "Fair_Use".}

\item{missing}{Maximum percentage of time steps allowed to be missing in any given year}

\item{gapfill_all}{Maximum percentage of time steps allowed to be gap-filled
(any quality) in any given year. Note if gapfill_all is set, any thresholds
for gapfill_good, gapfill_med or gapfill_poor are ignored. Set to NA if not required.}

\item{gapfill_good}{Maximum percentage of time steps allowed to be good-quality gap-filled
in any given year. Refer to package documentation for information on QC flags.
Set to NA if not required (default).}

\item{gapfill_med}{Maximum percentage of time steps allowed to be medium-quality gap-filled
in any given year. Refer to package documentation for information on QC flags.
Set to NA if not required (default).}

\item{gapfill_poor}{Maximum percentage of time steps allowed to be poor-quality gap-filled
in any given year. Refer to package documentation for information on QC flags.
Set to NA if not required (default).}

\item{min_yrs}{Minimum number of consecutive years to process}

\item{include_all_eval}{Should all evaluation values be included, regardless of data gaps?
If set to FALSE, any evaluation variables with missing or gap-filled values in
excess of the thresholds will be discarded.}

\item{plot}{Should annual, diurnal and/or 14-day running mean plots be produced?
       Set to NA if not required.}

\item{era_file}{ERA input file (needed if using ERAinterim to gapfill met variables)
e.g. "FULLSET/FLX_AU-How_FLUXNET2015_ERAI_HH_1989-2014_1-3.csv"}
}
\description{
Converts data from a FLUXNET2015 formatted spreadsheet to
NetCDF.
}
\details{
author: Anna Ukkola UNSW 2017.
Main function to convert Fluxnet2015 CSV-files to NetCDF
}

