% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FluxtowerSpreadsheetToNc.R
\name{convert_LaThuile}
\alias{convert_LaThuile}
\title{Converts La Thuile files to FLUXNET2015 format}
\usage{
convert_LaThuile(infiles, fair_usage = NA, fair_usage_vec = NA, min_yrs,
  tcol, time, site_log)
}
\description{
Converts La Thuile files to FLUXNET2015 format
}

