% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Handlers_csv.R
\name{convert_LaThuile}
\alias{convert_LaThuile}
\title{Converts La Thuile files to FLUXNET2015 format}
\usage{
convert_LaThuile(
  infiles,
  fair_usage = NA,
  fair_usage_vec = NA,
  min_yrs,
  tcol,
  add_psurf,
  site_log,
  site_code
)
}
\arguments{
\item{infiles}{input file}

\item{fair_usage}{fair use?}

\item{fair_usage_vec}{fair use?}

\item{min_yrs}{min year}

\item{tcol}{tcol?}

\item{add_psurf}{psurf?}

\item{site_log}{log file}

\item{site_code}{site code}
}
\value{
converted file in FLUXNET format
}
\description{
Converts La Thuile files to FLUXNET2015 format
}
