% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Handlers_csv.R
\name{convert_LaThuile}
\alias{convert_LaThuile}
\title{Converts La Thuile files to FLUXNET2015 format}
\usage{
convert_LaThuile(infiles, fair_usage = NA, fair_usage_vec = NA, min_yrs,
  tcol, add_psurf, site_log, site_code)
}
\description{
Converts La Thuile files to FLUXNET2015 format
}
