% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Check_and_Gapfill.R
\name{create_qc_var}
\alias{create_qc_var}
\title{Creates attributes for a new QC variable}
\usage{
create_qc_var(datain, qc_name)
}
\value{
datain
}
\description{
Creates attributes for a new QC variable
}

