% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FluxtowerSpreadsheetToNc.R
\name{CreateFluxNcFile}
\alias{CreateFluxNcFile}
\title{Creates a netcdf file for flux variables}
\usage{
CreateFluxNcFile(fluxfilename, datain, latitude, longitude, site_code,
  long_sitename, datasetversion, github_rev, tier = NA, ind_start, ind_end,
  starttime, timestepsize, flux_varname, cf_name, elevation = NA,
  towerheight = NA, canopyheight = NA, short_veg_type = NA,
  long_veg_type = NA, missing, gapfill_all, gapfill_good, gapfill_med,
  gapfill_poor, min_yrs, total_missing, total_gapfilled, QCmeasured,
  QCgapfilled, infile, var_ind)
}
\description{
Creates a netcdf file for flux variables
}

