% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Handlers_NetCDF.R
\name{CreateFluxNetcdfFile}
\alias{CreateFluxNetcdfFile}
\title{Creates a netcdf file for flux variables}
\usage{
CreateFluxNetcdfFile(
  fluxfilename,
  datain,
  site_code,
  siteInfo,
  ind_start,
  ind_end,
  starttime,
  flux_varname,
  cf_name,
  total_missing,
  total_gapfilled,
  qcInfo,
  arg_info,
  var_ind,
  varnames,
  modelInfo,
  global_atts
)
}
\arguments{
\item{fluxfilename}{output file name}

\item{datain}{input data}

\item{site_code}{site code}

\item{siteInfo}{meta-data}

\item{ind_start}{start year}

\item{ind_end}{end year}

\item{starttime}{timing info}

\item{flux_varname}{flux variable names}

\item{cf_name}{cf name}

\item{total_missing}{percentage missing and gap filled}

\item{total_gapfilled}{percentage missing and gap filled}

\item{qcInfo}{qc flag information}

\item{arg_info}{processing information}

\item{var_ind}{variable indices?}

\item{varnames}{original varnames corresponding to the dataset}

\item{modelInfo}{model parameters}

\item{global_atts}{global attributes (from origianl Ozflux nc-files)}
}
\value{
netcdf file with flux variables
}
\description{
Creates a netcdf file for flux variables
}
