% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FluxtowerSpreadsheetToNc.R
\name{CreateMetNcFile}
\alias{CreateMetNcFile}
\title{Creates a netcdf file for met variables}
\usage{
CreateMetNcFile(metfilename, datain, latitude, longitude, site_code,
  long_sitename, datasetversion, github_rev, tier, ind_start, ind_end,
  starttime, timestepsize, flux_varname, cf_name, elevation = NA,
  towerheight = NA, canopyheight = NA, short_veg_type = NA,
  long_veg_type = NA)
}
\description{
Creates a netcdf file for met variables
}
