% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Handlers_NetCDF.R
\name{CreateMetNetcdfFile}
\alias{CreateMetNetcdfFile}
\title{Creates a netcdf file for met variables}
\usage{
CreateMetNetcdfFile(
  metfilename,
  datain,
  site_code,
  siteInfo,
  ind_start,
  ind_end,
  starttime,
  flux_varname,
  cf_name,
  av_precip = NA,
  total_missing,
  total_gapfilled,
  qcInfo,
  qc_name,
  arg_info,
  var_ind,
  varnames,
  modelInfo,
  global_atts
)
}
\arguments{
\item{metfilename}{}

\item{datain}{}

\item{site_code}{}

\item{siteInfo}{}

\item{ind_start}{}

\item{ind_end}{}

\item{starttime}{}

\item{flux_varname}{}

\item{cf_name}{}

\item{av_precip}{}

\item{total_missing}{}

\item{total_gapfilled}{}

\item{qcInfo}{}

\item{qc_name}{}

\item{arg_info}{}

\item{var_ind}{}

\item{varnames}{}

\item{modelInfo}{}

\item{global_atts}{}
}
\description{
Creates a netcdf file for met variables
}
