% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Handlers_NetCDF.R
\name{define_model_params}
\alias{define_model_params}
\title{Writes model parameters as global attribute to NetCDF file}
\usage{
define_model_params(modelInfo, dims)
}
\arguments{
\item{modelInfo}{model parameters}

\item{dims}{dimensions}
}
\description{
Writes model parameters as global attribute to NetCDF file
}
