% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Constants.R
\name{duplicate_columns}
\alias{duplicate_columns}
\title{Duplicates columns in Fluxnet data}
\usage{
duplicate_columns(data, vars)
}
\arguments{
\item{data}{input data}

\item{vars}{variables}
}
\value{

}
\description{
If a variable is being processes multiple times duplicate column
}
