% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Constants.R
\name{duplicate_columns}
\alias{duplicate_columns}
\title{Duplicates columns in Fluxnet data if a variable is being processes multiple times}
\usage{
duplicate_columns(data, vars)
}
\value{
data with duplicated columns
}
\description{
Duplicates columns in Fluxnet data if a variable is being processes multiple times
}
