% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Gapfilling.R
\name{FillQCvarMissing}
\alias{FillQCvarMissing}
\title{Fills QC flags with 3 (poor gap-filling) when
QC flag missing but data variable available}
\usage{
FillQCvarMissing(datain, gapfillVal, qc_name)
}
\value{
datain
}
\description{
Fills QC flags with 3 (poor gap-filling) when
QC flag missing but data variable available
}

