% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Check_and_Gapfill.R
\name{find_ind_and_qc}
\alias{find_ind_and_qc}
\title{Find index for a variable and its QC flag}
\usage{
find_ind_and_qc(inds, var, qc_name = NA)
}
\description{
Find index for a variable and its QC flag
}

