% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Constants.R
\name{findColIndices}
\alias{findColIndices}
\title{Finds variables present in input file}
\usage{
findColIndices(fileinname, var_names, var_classes, essential_vars,
  preferred_vars, time_vars)
}
\value{
list of variables and their attributes
}
\description{
Finds variables present in input file
}

