% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Constants.R
\name{findColIndices}
\alias{findColIndices}
\title{Finds variables present in input file}
\usage{
findColIndices(fileinname, var_names, var_classes, essential_vars,
  preferred_vars, time_vars, dset_vars, site_log, datasetname, ...)
}
\value{
list of variables and their attributes
}
\description{
Finds variables present in input file
}
