% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gapfilling.R
\name{FindFluxInd}
\alias{FindFluxInd}
\title{Finds indices for flux variables to be outputted}
\usage{
FindFluxInd(datain, exclude_eval, k, site_log)
}
\description{
Finds indices for flux variables to be outputted
}
