% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gapfilling.R
\name{GapfillFlux}
\alias{GapfillFlux}
\title{Gapfill flux variables using statistical methods}
\usage{
GapfillFlux(
  datain,
  qc_name,
  qc_flags,
  regfill,
  linfill,
  copyfill,
  varnames,
  site_log
)
}
\arguments{
\item{datain}{}

\item{qc_name}{}

\item{qc_flags}{}

\item{regfill}{}

\item{linfill}{}

\item{copyfill}{}

\item{varnames}{}

\item{site_log}{}
}
\value{
gapfilled flux variables
}
\description{
Gapfill flux variables using statistical methods
}
