% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gapfilling.R
\name{GapfillFlux}
\alias{GapfillFlux}
\title{Gapfill flux variables using statistical methods}
\usage{
GapfillFlux(datain, qc_name, qc_flags, regfill, linfill, copyfill,
  varnames, site_log)
}
\description{
Gapfill flux variables using statistical methods
}
