% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gapfilling.R
\name{GapfillMet_statistical}
\alias{GapfillMet_statistical}
\title{Gapfill meteorological variables using statistical methods}
\usage{
GapfillMet_statistical(datain, qc_name, qc_flags, copyfill, linfill,
  lwdown_method, elevation, varnames, site_log)
}
\description{
Gapfill meteorological variables using statistical methods
}
