% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gapfilling.R
\name{GapfillMet_statistical}
\alias{GapfillMet_statistical}
\title{Gapfill meteorological variables using statistical methods}
\usage{
GapfillMet_statistical(
  datain,
  qc_name,
  qc_flags,
  copyfill,
  linfill,
  lwdown_method,
  elevation,
  varnames,
  limit_vars,
  site_log
)
}
\arguments{
\item{datain}{input data}

\item{qc_name}{qc name}

\item{qc_flags}{qc flag?}

\item{copyfill}{copyfill?}

\item{linfill}{linear interpollation?}

\item{lwdown_method}{long wave radiation method}

\item{elevation}{elevation}

\item{varnames}{varnames?}

\item{limit_vars}{limits vars?}

\item{site_log}{log file}
}
\description{
Gapfill meteorological variables using statistical methods
}
