% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Gapfilling.R
\name{GapfillMet_with_ERA}
\alias{GapfillMet_with_ERA}
\title{Gapfills meteorological data with down-scaled ERAinterim estimates}
\usage{
GapfillMet_with_ERA(datain, ERA_file, qc_name, varnames, ...)
}
\description{
Gapfills meteorological data with down-scaled ERAinterim estimates
}

