% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helpers.R
\name{get_fluxnet_file_template}
\alias{get_fluxnet_file_template}
\title{Gets a Fluxnet file template (or full file name if site codes and years are specified)}
\usage{
get_fluxnet_file_template(site_code = "[A-Z]{2}-[A-Za-z0-9]{3}",
  datasetname = "FLUXNET2015", subset = "FULLSET",
  resolution = "[A-Z]{2}", flx2015_years = "[0-9]{4}-[0-9]{4}",
  lathuile_year = "[0-9]{4}", datasetversion = "[0-9]{1}-[0-9]{1}",
  extension = ".csv")
}
\description{
Gets a Fluxnet file template (or full file name if site codes and years are specified)
}

