% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Helpers.R
\name{get_fluxnet_version_no}
\alias{get_fluxnet_version_no}
\title{Gets Fluxnet dataset version from Fluxnet file}
\usage{
get_fluxnet_version_no(file)
}
\description{
Gets Fluxnet dataset version from Fluxnet file
}

