% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Checks.R
\name{InitialChecks}
\alias{InitialChecks}
\title{Performs initial checks on function arguments}
\usage{
InitialChecks(met_gapfill, era_file, missing, aggregate, datasetname,
  flx2015_version)
}
\description{
Performs initial checks on function arguments
}

