% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Plotting.R
\name{plot_nc}
\alias{plot_nc}
\title{Plots standard analysis plots from netcdf data}
\usage{
plot_nc(ncfile, analysis_type, vars, outfile)
}
\arguments{
\item{ncfile}{an open netcdf file}

\item{analysis_type}{vector of plot names: c("annual", "diurnal", "timeseries")}

\item{vars}{vector of variable names to plot}

\item{outfile}{output path prefix, including directory}
}
\description{
Plots standard analysis plots from netcdf data
}

