% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Aggregation.R
\name{qc_frac}
\alias{qc_frac}
\title{Calculates fraction of good quality gapfilled and observed data
for aggregated time steps}
\usage{
qc_frac(data, good_data)
}
\description{
Calculates fraction of good quality gapfilled and observed data
for aggregated time steps
}
