% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FluxtowerSpreadsheetToNc.R
\name{ReadTextFluxData}
\alias{ReadTextFluxData}
\title{Reads comma-delimited text files containing Fluxnet2015 data}
\usage{
ReadTextFluxData(fileinname, vars, time_vars)
}
\value{
filedata
}
\description{
Reads comma-delimited text files containing Fluxnet2015 data
}

