% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Synthesize_and_gapfill_functions.R
\name{regfill_flux}
\alias{regfill_flux}
\title{Gapfills flux data using linear regression against met variables}
\usage{
regfill_flux(ydata, traindata, tstepsize, regfill, varname, swdown_ind,
  tair_ind, rh_ind, site_log, ...)
}
\description{
Gapfills flux data using linear regression against met variables
}

