% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Constants.R
\name{rename_vars}
\alias{rename_vars}
\title{Renames Fluxnet variables to ALMA convention}
\usage{
rename_vars(vars_present, all_vars)
}
\value{
renamed_vars
}
\description{
Renames Fluxnet variables to ALMA convention
}

