% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Constants.R
\name{retrieve_atts}
\alias{retrieve_atts}
\title{Retrieve variables longnames to be written in NetCDF}
\usage{
retrieve_atts(vars_present, all_vars)
}
\arguments{
\item{vars_present}{current variables?}

\item{all_vars}{all variables?}
}
\value{

}
\description{
Retrieve variables longnames to be written in NetCDF
}
