% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Constants.R
\name{retrieve_ERAvars}
\alias{retrieve_ERAvars}
\title{Retrieves names of ERAinterim variables}
\usage{
retrieve_ERAvars(vars_present, all_vars)
}
\value{
cat_vars
}
\description{
Retrieves names of ERAinterim variables
}

