% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Constants.R
\name{retrieve_ranges}
\alias{retrieve_ranges}
\title{Retrieves acceptable variable ranges}
\usage{
retrieve_ranges(vars_present, all_vars)
}
\value{
variable ranges
}
\description{
Retrieves acceptable variable ranges
}

