% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Constants.R
\name{retrieve_varinfo}
\alias{retrieve_varinfo}
\title{Retrieves variable information}
\usage{
retrieve_varinfo(vars_present, all_vars, attribute)
}
\value{
variable information
}
\description{
Retrieves variable information
}

