% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gapfilling.R
\name{update_qc}
\alias{update_qc}
\title{Updates QC flags after gap-filling}
\usage{
update_qc(data, temp_data, varname, qc_name, qc_value, qc_flags, ...)
}
\description{
Updates QC flags after gap-filling
}
