% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Conversions.R
\name{VPD2RelHum}
\alias{VPD2RelHum}
\title{Converts VPD (hPa) to relative humidity (percentage)}
\usage{
VPD2RelHum(VPD, airtemp, vpd_units, tair_units, site_log)
}
\value{
relative humidity as percentage
}
\description{
Converts VPD (hPa) to relative humidity (percentage)
}
